# Conflicts:
#	apps/task/urls.py

+++ apps/task/views.py
@@ -4,7 +4,7 @@ from drf_yasg.utils import swagger_auto_schema
 from rest_framework import viewsets
 from rest_framework.generics import GenericAPIView, get_object_or_404
 from rest_framework.response import Response
-from apps.task.serializers import FilterTaskSerializer, TaskUpdateAllSerializer
+from apps.task.serializers import TaskUpdateAllSerializer, TaskSearchSerializer
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.status import HTTP_204_NO_CONTENT
 from apps.task.serializers import TaskSelfSerializer
@@ -15,6 +15,7 @@ from apps.notification.views import AddNotificationTask, AddNotificationTaskClos
 from apps.users.serializers import UserTaskSerializer
 from django.contrib.auth.models import User
 from rest_framework.pagination import PageNumberPagination
+from rest_framework import filters
 
 
 class TenResultsSetPagination(PageNumberPagination):
@@ -37,7 +38,7 @@ class TaskFilterStatusCreatedViewSet(viewsets.ModelViewSet):
     authentication_classes = ()
 
     serializer_class = TaskSerializer
-    queryset = Task.objects.filter(status=Task.CREATED)
+    queryset = Task.objects.filter(status=Task.CREATED).order_by('-id')
 
     pagination_class = TenResultsSetPagination
     http_method_names = ['get']
@@ -48,7 +49,7 @@ class TaskFilterStatusInprocessViewSet(viewsets.ModelViewSet):
     authentication_classes = ()
 
     serializer_class = TaskSerializer
-    queryset = Task.objects.filter(status=Task.INPROCESS)
+    queryset = Task.objects.filter(status=Task.INPROCESS).order_by('-id')
 
     pagination_class = TenResultsSetPagination
     http_method_names = ['get']
@@ -59,7 +60,7 @@ class TaskFilterStatusFinishedViewSet(viewsets.ModelViewSet):
     authentication_classes = ()
 
     serializer_class = TaskSerializer
-    queryset = Task.objects.filter(status=Task.FINISHED)
+    queryset = Task.objects.filter(status=Task.FINISHED).order_by('-id')
 
     pagination_class = TenResultsSetPagination
     http_method_names = ['get']
@@ -112,7 +113,9 @@ class DeleteView(GenericAPIView):
     permission_classes = (IsAuthenticated,)
 
     def delete(self, request, pk):
-        task = get_object_or_404(Task.objects.filter(pk=pk))
+        task = Task.objects.filter(pk=pk, user_created=request.user.id)
+        if task.count() == 0:
+            return Response(status=403)
         task.delete()
         return Response(status=HTTP_204_NO_CONTENT)
 
@@ -258,3 +261,15 @@ class UpdateTask(GenericAPIView):
             return Response(response_data)
         else:
             return Response(serializer.errors, status=400)
+
+
+class TaskSearchViewSet(viewsets.ModelViewSet):
+    permission_classes = (AllowAny,)
+    authentication_classes = ()
+
+    serializer_class = TaskSearchSerializer
+    queryset = Task.objects.all()
+
+    filter_backends = (filters.SearchFilter,)
+    search_fields = ('$title', '$description')
+    http_method_names = ['get']
